Add simple caching for meta queries
